refactor(plugin structure): move classes into namespaces
