+++ includes/Options/FulfillmentOptionsTab.php
@@ -5,9 +5,11 @@
  * @package Options
  */
 
-namespace Towa\GebruederWeissWooCommerce;
+namespace Towa\GebruederWeissWooCommerce\Options;
 
 defined('ABSPATH') || exit;
+
+use Towa\GebruederWeissWooCommerce\GbWeiss;
 
 /**
  * Fulfillment Options Tab
